+++ app/api/v1/keys.py
@@ -6,6 +6,7 @@ from chia.rpc.wallet_rpc_client import WalletRpcClient
 from chia.types.blockchain_format.sized_bytes import bytes32
 from chia.util.bech32m import decode_puzzle_hash, encode_puzzle_hash
 from chia.util.byte_types import hexstr_to_bytes
+from chia.util.ints import uint32
 from chia.wallet.derive_keys import (
     master_sk_to_wallet_sk,
     master_sk_to_wallet_sk_unhardened,
@@ -31,7 +32,6 @@ async def get_key(
     prefix: str = "bls1238",
     wallet_rpc_client: WalletRpcClient = Depends(deps.get_wallet_rpc_client),
 ):
-
     fingerprint: int = await wallet_rpc_client.get_logged_in_fingerprint()
 
     result: Dict = await wallet_rpc_client.get_private_key(fingerprint)
@@ -40,10 +40,10 @@ async def get_key(
 
     wallet_secret_key: PrivateKey
     if hardened:
-        wallet_secret_key = master_sk_to_wallet_sk(secret_key, derivation_index)
+        wallet_secret_key = master_sk_to_wallet_sk(secret_key, uint32(derivation_index))
     else:
         wallet_secret_key = master_sk_to_wallet_sk_unhardened(
-            secret_key, derivation_index
+            secret_key, uint32(derivation_index)
         )
 
     wallet_public_key: G1Element = wallet_secret_key.get_g1()
@@ -63,7 +63,6 @@ async def get_key(
 async def parse_key(
     address: str,
 ):
-
     try:
         puzzle_hash: bytes = decode_puzzle_hash(address)
     except ValueError: