+++ app/models/__init__.py
@@ -1,2 +1,4 @@
 from app.models.activity import Activity  # noqa
 from app.models.state import State  # noqa
+
+__all__ = ["Activity", "State"]