+++ tests/test_cat_workflow.py
@@ -1,14 +1,12 @@
 import logging
-from typing import Dict, List
+from typing import List
 
 import pytest
 from blspy import PrivateKey
-from chia.consensus.constants import ConsensusConstants
 from chia.rpc.wallet_rpc_client import WalletRpcClient
-from chia.server.server import ChiaServer
 from chia.simulator.full_node_simulator import FullNodeSimulator
+from chia.simulator.time_out_assert import time_out_assert
 from chia.types.blockchain_format.sized_bytes import bytes32
-from chia.types.spend_bundle import SpendBundle
 from chia.wallet.transaction_record import TransactionRecord
 from chia.wallet.wallet import Wallet
 from chia.wallet.wallet_node import WalletNode
@@ -16,12 +14,10 @@ from chia.wallet.wallet_node import WalletNode
 from app.core.climate_wallet.wallet import ClimateObserverWallet, ClimateWallet
 from app.core.derive_keys import master_sk_to_root_sk
 from app.core.types import ClimateTokenIndex, GatewayMode
-from app.core.utils import get_first_puzzle_hash
-from tests.wallet.rpc.test_wallet_rpc import wallet_rpc_environment  # noqa
+from tests.wallet.rpc.test_wallet_rpc import wallet_rpc_environment  # noqa: F401
 from tests.wallet.rpc.test_wallet_rpc import (
     WalletRpcTestEnvironment,
     farm_transaction,
-    farm_transaction_block,
     generate_funds,
 )
 
@@ -33,7 +29,6 @@ async def check_transactions(
     wallet_id: int,
     transaction_records: List[TransactionRecord],
 ) -> None:
-
     for transaction_record in transaction_records:
         tx = await wallet_client.get_transaction(
             wallet_id=wallet_id, transaction_id=transaction_record.name
@@ -41,7 +36,7 @@ async def check_transactions(
 
         assert (
             tx.confirmed_at_height != 0
-        ), "Transaction {transaction_record.name.hex()} not found!"
+        ), f"Transaction {transaction_record.name.hex()} not found!"
 
 
 async def check_balance(
@@ -49,11 +44,14 @@ async def check_balance(
     wallet_id: int,
     amount: int,
 ) -> None:
-
-    result: Dict = await wallet_client.get_wallet_balance(wallet_id=wallet_id)
+    result = await wallet_client.get_wallet_balance(wallet_id=wallet_id)
     assert (
         result["confirmed_wallet_balance"] == amount
     ), "Target wallet CAT amount does not match!"
+
+
+async def get_confirmed_balance(client: WalletRpcClient, wallet_id: int) -> int:
+    return int((await client.get_wallet_balance(wallet_id))["confirmed_wallet_balance"])
 
 
 class TestCATWorkflow:
@@ -69,15 +67,14 @@ class TestCATWorkflow:
     @pytest.mark.asyncio
     async def test_cat_tokenization_workflow(
         self,
-        wallet_rpc_environment: WalletRpcTestEnvironment,
+        wallet_rpc_environment: WalletRpcTestEnvironment,  # noqa: F811
         org_uid: str,
         warehouse_project_id: str,
         vintage_year: int,
         amount: int,
         fee: int,
     ) -> None:
-
-        env: WalletRpcTestEnvironment = wallet_rpc_environment
+        env = wallet_rpc_environment
 
         wallet_node_1: WalletNode = env.wallet_1.node
         wallet_client_1: WalletRpcClient = env.wallet_1.rpc_client
@@ -88,7 +85,7 @@ class TestCATWorkflow:
         full_node_api: FullNodeSimulator = env.full_node.api
 
         fingerprint: int = await wallet_client_1.get_logged_in_fingerprint()
-        result: Dict = await wallet_client_1.get_private_key(fingerprint=fingerprint)
+        result = await wallet_client_1.get_private_key(fingerprint=fingerprint)
         master_secret_key: PrivateKey = PrivateKey.from_bytes(
             bytes.fromhex(result["sk"])
         )
@@ -113,39 +110,39 @@ class TestCATWorkflow:
             root_secret_key=root_secret_key,
             wallet_client=wallet_client_1,
         )
-        result: Dict = await climate_wallet_1.send_tokenization_transaction(
+        result = await climate_wallet_1.send_tokenization_transaction(
             to_puzzle_hash=await wallet_2.get_new_puzzlehash(),
             amount=amount,
             fee=fee,
         )
-        spend_bundle: SpendBundle = result["spend_bundle"]
         transaction_records: List[TransactionRecord] = result["transaction_records"]
 
-        await farm_transaction(full_node_api, wallet_node_1, spend_bundle)
+        await full_node_api.process_all_wallet_transactions(
+            wallet=wallet_node_1.wallet_state_manager.main_wallet, timeout=120
+        )
         await check_transactions(wallet_client_1, 1, transaction_records)
 
         # block:
         #   - client: create CAT wallet
 
-        result: List[Dict] = await wallet_client_2.get_stray_cats()
+        result = await wallet_client_2.get_stray_cats()
         asset_id = bytes32.fromhex(result[0]["asset_id"])
-        result: Dict = await wallet_client_2.create_wallet_for_existing_cat(
-            asset_id=asset_id
-        )
+        result = await wallet_client_2.create_wallet_for_existing_cat(asset_id=asset_id)
+        assert result["success"]
         cat_wallet_id: int = result["wallet_id"]
 
-        await farm_transaction_block(full_node_api, wallet_node_1)
-        await check_balance(wallet_client_2, cat_wallet_id, amount)
+        await time_out_assert(
+            60, get_confirmed_balance, amount, wallet_client_2, cat_wallet_id
+        )
 
     @pytest.mark.asyncio
     async def test_cat_detokenization_workflow(
         self,
-        wallet_rpc_environment: WalletRpcTestEnvironment,
+        wallet_rpc_environment: WalletRpcTestEnvironment,  # noqa: F811
         token_index: ClimateTokenIndex,
         amount: int = 10,
         fee: int = 10,
     ) -> None:
-
         env: WalletRpcTestEnvironment = wallet_rpc_environment
 
         wallet_node_1: WalletNode = env.wallet_1.node
@@ -157,12 +154,11 @@ class TestCATWorkflow:
         full_node_api: FullNodeSimulator = env.full_node.api
 
         fingerprint: int = await wallet_client_1.get_logged_in_fingerprint()
-        result: Dict = await wallet_client_1.get_private_key(fingerprint=fingerprint)
+        result = await wallet_client_1.get_private_key(fingerprint=fingerprint)
         master_secret_key: PrivateKey = PrivateKey.from_bytes(
             bytes.fromhex(result["sk"])
         )
         root_secret_key: PrivateKey = master_sk_to_root_sk(master_secret_key)
-        constants: ConsensusConstants = full_node_api.full_node.constants
 
         # block: initial fund deposits
 
@@ -178,20 +174,22 @@ class TestCATWorkflow:
             root_secret_key=root_secret_key,
             wallet_client=wallet_client_1,
         )
-        result: Dict = await climate_wallet_1.send_tokenization_transaction(
+        result = await climate_wallet_1.send_tokenization_transaction(
             to_puzzle_hash=await wallet_2.get_new_puzzlehash(),
             amount=amount,
             fee=fee,
         )
-        spend_bundle: SpendBundle = result["spend_bundle"]
+        # spend_bundle: SpendBundle = result["spend_bundle"]
         transaction_records: List[TransactionRecord] = result["transaction_records"]
 
-        result: Dict = await wallet_client_2.create_wallet_for_existing_cat(
+        result = await wallet_client_2.create_wallet_for_existing_cat(
             asset_id=climate_wallet_1.tail_program.get_tree_hash()
         )
         cat_wallet_id: int = result["wallet_id"]
 
-        await farm_transaction(full_node_api, wallet_node_1, spend_bundle)
+        await full_node_api.process_all_wallet_transactions(
+            wallet=wallet_node_1.wallet_state_manager.main_wallet, timeout=120
+        )
         await check_transactions(wallet_client_1, 1, transaction_records)
 
         # block:
@@ -207,15 +205,15 @@ class TestCATWorkflow:
             wallet_client=wallet_client_2,
             constants=climate_wallet_1.constants,
         )
-        result: Dict = await climate_wallet_2.create_detokenization_request(
+        result = await climate_wallet_2.create_detokenization_request(
             amount=amount,
             fee=fee,
             wallet_id=cat_wallet_id,
         )
         content: str = result["content"]
-        transaction_records: List[TransactionRecord] = result["transaction_records"]
+        transaction_records = result["transaction_records"]
 
-        result: Dict = await ClimateWallet.parse_detokenization_request(
+        result = await ClimateWallet.parse_detokenization_request(
             content=content,
         )
         assert result["mode"] == GatewayMode.DETOKENIZATION
@@ -223,44 +221,40 @@ class TestCATWorkflow:
         assert result["fee"] == fee
         assert result["asset_id"] == climate_wallet_1.tail_program_hash
 
-        result: Dict = await climate_wallet_1.sign_and_send_detokenization_request(
+        result = await climate_wallet_1.sign_and_send_detokenization_request(
             content=content,
         )
-        spend_bundle: SpendBundle = result["spend_bundle"]
+        spend_bundle = result["spend_bundle"]
 
         await farm_transaction(full_node_api, wallet_node_1, spend_bundle)
+        await full_node_api.wait_for_wallet_synced(env.wallet_2.node, timeout=60)
         await check_transactions(wallet_client_2, cat_wallet_id, transaction_records)
-        await check_balance(wallet_client_2, cat_wallet_id, 0)
+        await time_out_assert(
+            60, get_confirmed_balance, 0, wallet_client_2, cat_wallet_id
+        )
 
     @pytest.mark.asyncio
     async def test_cat_permissionless_retirement_workflow(
         self,
-        wallet_rpc_environment: WalletRpcTestEnvironment,
+        wallet_rpc_environment: WalletRpcTestEnvironment,  # noqa: F811
         token_index: ClimateTokenIndex,
         amount: int = 10,
         fee: int = 10,
-        beneficiary_name: bytes = b"Ionia",
+        beneficiary_name: bytes = "Ionia".encode(),
     ) -> None:
-
         env: WalletRpcTestEnvironment = wallet_rpc_environment
 
-        wallet_node_1: WalletNode = env.wallet_1.node
         wallet_client_1: WalletRpcClient = env.wallet_1.rpc_client
-
-        wallet_node_2: WalletNode = env.wallet_2.node
         wallet_client_2: WalletRpcClient = env.wallet_2.rpc_client
         wallet_2: Wallet = env.wallet_2.wallet
 
         full_node_api: FullNodeSimulator = env.full_node.api
-        full_node_client: ChiaServer = env.full_node.rpc_client
+        full_node_client = env.full_node.rpc_client
 
         fingerprint: int = await wallet_client_1.get_logged_in_fingerprint()
-        result: Dict = await wallet_client_1.get_private_key(fingerprint=fingerprint)
-        master_secret_key: PrivateKey = PrivateKey.from_bytes(
-            bytes.fromhex(result["sk"])
-        )
-        root_secret_key: PrivateKey = master_sk_to_root_sk(master_secret_key)
-        constants: ConsensusConstants = full_node_api.full_node.constants
+        result = await wallet_client_1.get_private_key(fingerprint=fingerprint)
+        master_secret_key = PrivateKey.from_bytes(bytes.fromhex(result["sk"]))
+        root_secret_key = master_sk_to_root_sk(master_secret_key)
 
         # block: initial fund deposits
 
@@ -276,20 +270,21 @@ class TestCATWorkflow:
             root_secret_key=root_secret_key,
             wallet_client=wallet_client_1,
         )
-        result: Dict = await climate_wallet_1.send_tokenization_transaction(
+        result = await climate_wallet_1.send_tokenization_transaction(
             to_puzzle_hash=await wallet_2.get_new_puzzlehash(),
             amount=amount,
             fee=fee,
         )
-        spend_bundle: SpendBundle = result["spend_bundle"]
-        transaction_records: List[TransactionRecord] = result["transaction_records"]
+        transaction_records = result["transaction_records"]
 
-        result: Dict = await wallet_client_2.create_wallet_for_existing_cat(
+        result = await wallet_client_2.create_wallet_for_existing_cat(
             asset_id=climate_wallet_1.tail_program.get_tree_hash()
         )
         cat_wallet_id: int = result["wallet_id"]
 
-        await farm_transaction(full_node_api, wallet_node_1, spend_bundle)
+        await full_node_api.process_all_wallet_transactions(
+            wallet=env.wallet_1.node.wallet_state_manager.main_wallet, timeout=120
+        )
         await check_transactions(wallet_client_1, 1, transaction_records)
 
         # block:
@@ -303,20 +298,23 @@ class TestCATWorkflow:
             wallet_client=wallet_client_2,
             constants=climate_wallet_1.constants,
         )
-        result: Dict = (
-            await climate_wallet_2.send_permissionless_retirement_transaction(
-                amount=amount,
-                fee=fee,
-                beneficiary_name=beneficiary_name,
-                wallet_id=cat_wallet_id,
-            )
-        )
-        spend_bundle: SpendBundle = result["spend_bundle"]
-        transaction_records: List[TransactionRecord] = result["transaction_records"]
 
-        await farm_transaction(full_node_api, wallet_node_2, spend_bundle)
-        await check_transactions(wallet_client_2, cat_wallet_id, transaction_records)
-        await check_balance(wallet_client_2, cat_wallet_id, 0)
+        test_address = "This is a fake address".encode()
+        result = await climate_wallet_2.send_permissionless_retirement_transaction(
+            amount=amount,
+            fee=fee,
+            beneficiary_name=beneficiary_name,
+            beneficiary_address=test_address,
+            wallet_id=cat_wallet_id,
+        )
+        transaction_records = result["transaction_records"]
+
+        await full_node_api.process_all_wallet_transactions(
+            wallet=env.wallet_2.node.wallet_state_manager.main_wallet, timeout=120
+        )
+        await time_out_assert(
+            60, get_confirmed_balance, 0, wallet_client_2, cat_wallet_id
+        )
 
         # block:
         #   - observer: observe retirement activity
@@ -326,12 +324,10 @@ class TestCATWorkflow:
             root_public_key=climate_wallet_1.root_public_key,
             full_node_client=full_node_client,
         )
-        activities: List[Dict] = await climate_observer.get_activities(
+        activities = await climate_observer.get_activities(
             mode=GatewayMode.PERMISSIONLESS_RETIREMENT
         )
 
         assert activities[0]["metadata"]["bn"] == beneficiary_name.decode()
-        assert (
-            activities[0]["metadata"]["bp"]
-            == "0x" + (await get_first_puzzle_hash(wallet_client_2)).hex()
-        )
+        assert activities[0]["metadata"]["ba"] == test_address.decode()
+        assert activities[0]["metadata"]["bp"] == "0x"