+++ tests/test_crud_core.py
@@ -4,12 +4,12 @@ from app.crud.db import DBCrud
 
 
 class TestUpdateBlockState:
-    def test_with_block_height_then_success(self):
+    def test_with_block_height_then_success(self) -> None:
         mock_db = mock.MagicMock()
         actual = DBCrud(db=mock_db).update_block_state(peak_height=1)
         assert actual is True
 
-    def test_with_current_height_then_success(self):
+    def test_with_current_height_then_success(self) -> None:
         mock_db = mock.MagicMock()
         actual = DBCrud(db=mock_db).update_block_state(current_height=1)
         assert actual is True