+++ CHANGELOG.md
@@ -1,3 +1,32 @@
+## 1.0.33 (2023-10-12)
+
+### Fix
+
+- add missing max coin amount in select coin call
+
+## 1.0.32 (2023-10-11)
+
+## 1.0.31 (2023-10-11)
+
+## 1.0.30 (2023-10-11)
+
+### Fix
+
+- derivation roots
+
+## 1.0.29 (2023-10-05)
+
+### Fix
+
+- update the default cadt api host
+
+## 1.0.28 (2023-10-03)
+
+### Fix
+
+- rework wallet indexes
+- update default CADT API in config
+
 ## 1.0.27 (2023-09-29)
 
 ### Fix